requirements.txt
streamlit
python-dotenv
pypdf
pandas
numpy
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-classic
langchain-huggingface
langchain-google-genai
faiss-cpu
transformers
sentence-transformers
torch
